+++ jovian/utils/commit.py
@@ -6,8 +6,7 @@ from jovian.utils.jupyter import in_notebook, get_notebook_name, save_notebook
 from jovian.utils.misc import get_file_extension, is_uuid
 from jovian.utils.rcfile import get_notebook_slug, set_notebook_slug
 from jovian.utils.credentials import read_webapp_url
-from jovian.utils.anaconda import upload_conda_env, CondaError
-from jovian.utils.pip import upload_pip_env
+from jovian.utils.environment import upload_conda_env, CondaError, upload_pip_env
 from jovian.utils.records import log_git, get_record_slugs, reset_records
 from jovian.utils.constants import FILENAME_MSG
 from jovian.utils.logger import log
