+++ jovian/utils/environment.py
@@ -1,4 +1,4 @@
-"""Anaconda related utilities"""
+"""Anaconda and Pip related utilities"""
 import os
 import logging
 from jovian.utils.api import upload_file
@@ -82,3 +82,18 @@ def print_conda_message(env_name):
 #     $ conda deactivate
         """) % env_name
         log(message)
+
+
+def read_pip_env():
+    """Read the pip dependencies into a string"""
+    command = "pip --disable-pip-version-check freeze"
+    deps_str = os.popen(command).read()
+    if deps_str == '':
+        error = 'Failed to read Anaconda environment using command: "' + command + '"'
+        raise Exception(error)
+    return deps_str
+
+
+def upload_pip_env(gist_slug, version=None):
+    """Read and upload the current virtual environment to server"""
+    return upload_file(gist_slug=gist_slug, file=('requirements.txt', read_pip_env()), version=version)
