+++ jovian/utils/misc.py
@@ -49,3 +49,13 @@ def is_flavor_pro():
 def get_file_extension(filename):
     """Get the extension of a file"""
     return filename.split('.')[-1] if type(filename) == str else ''
+
+
+def urljoin(*args):
+    """Join multiple url parts to construct one url"""
+    if len(args) == 0:
+        raise TypeError("urljoin requires at least one argument")
+
+    trailing_slash = '/' if args[-1].endswith('/') else ''
+
+    return '/'.join(map(lambda x: str(x).strip('/'), args)) + trailing_slash
